[CSL-1279] Create Pos.Ssc.Message
