+++ lib/src/Pos/Lrc/Consumers.hs
@@ -10,7 +10,7 @@ import           Pos.Delegation.Lrc    (delegationLrcConsumer)
 import           Pos.Lrc.Consumer      (LrcConsumer)
 import           Pos.Lrc.Mode          (LrcMode)
 import           Pos.Ssc.Lrc           (sscLrcConsumer)
-import           Pos.Ssc.Network.Constraint (GtMessageConstraints)
+import           Pos.Ssc.Message       (GtMessageConstraints)
 import           Pos.Update.Lrc        (usLrcConsumer)
 
 allLrcConsumers
