+++ ssc/Pos/Binary/Ssc/Relay.hs
@@ -7,7 +7,7 @@ import           Universum
 import           Pos.Binary.Class              (Bi (..), encodeListLen, enforceSize)
 import           Pos.Communication.Types.Relay (DataMsg (..))
 import           Pos.Core.Configuration        (HasConfiguration)
-import qualified Pos.Ssc.Types.Message         as T
+import qualified Pos.Ssc.Message               as T
 
 instance Bi (DataMsg T.MCCommitment) where
   encode (DataMsg (T.MCCommitment signedComm)) = encode signedComm
